Swtich the mason path modifier to append
This means that any tools installed out side of mason will take precedence. This
was necessary to solve a problem with mdformat, but will also be useful for
other tools. To change the behaviour of mdformat I have to install other python
packages but this seems to be beyond mason capabilities.